% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_as.R
\name{as_vlmc.ctx_tree}
\alias{as_vlmc.ctx_tree}
\title{Convert an object to a Variable Length Markov Chain (VLMC)}
\usage{
\method{as_vlmc}{ctx_tree}(x, ...)
}
\arguments{
\item{x}{an object to convert into a vlmc.}

\item{...}{additional arguments for conversion functions.}
}
\value{
a vlmc
}
\description{
This generic function converts an object into a vlmc.
}
\examples{
dts <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 0)
dts_ctree <- ctx_tree(dts, min_size = 1, max_depth = 3)
draw(dts_ctree)
dts_vlmc <- as_vlmc(dts_ctree)
class(dts_vlmc)
draw(dts_vlmc)
}
