% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_as.R
\name{as_vlmc}
\alias{as_vlmc}
\title{Convert an object to a Variable Length Markov Chain (VLMC)}
\usage{
as_vlmc(x, ...)
}
\arguments{
\item{x}{an object to convert into a vlmc.}

\item{...}{additional arguments for conversion functions.}
}
\value{
a vlmc
}
\description{
This generic function converts an object into a vlmc.
}
