% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_as.R
\name{as_vlmc.tune_vlmc}
\alias{as_vlmc.tune_vlmc}
\title{Convert an object to a Variable Length Markov Chain (VLMC)}
\usage{
\method{as_vlmc}{tune_vlmc}(x, ...)
}
\arguments{
\item{x}{an object to convert into a vlmc.}

\item{...}{additional arguments for conversion functions.}
}
\value{
a vlmc
}
\description{
This generic function converts an object into a vlmc.
}
\examples{
dts <- sample(as.factor(c("A", "B", "C")), 100, replace = TRUE)
tune_result <- tune_vlmc(dts)
tune_result
dts_best_vlmc <- as_vlmc(tune_result)
draw(dts_best_vlmc)
}
