% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree_draw.R
\name{charset_ascii}
\alias{charset_ascii}
\title{ASCII character set for context tree text representation}
\usage{
charset_ascii(
  root = "*",
  first_node = "+",
  next_node = "'",
  final_node = "'",
  vbranch = "|",
  hbranch = "--",
  open_ct = "(",
  close_ct = ")",
  level_sep = " ~ ",
  time_sep = " | ",
  intercept = "(I)",
  intercept_sep = " & ",
  open_p_value = "<",
  close_p_value = ">",
  open_model = "[",
  close_model = "]"
)
}
\arguments{
\item{root}{character used for the root node.}

\item{first_node}{characters used for the first child of a node.}

\item{next_node}{characters used for intermediate children of a node.}

\item{final_node}{characters used for the last child of a node.}

\item{vbranch}{characters used to represent a branch in a vertical way.}

\item{hbranch}{characters used to represent a branch in a horizontal was.}

\item{open_ct}{characters used to start each node specific text
representation.}

\item{close_ct}{characters used to end each node specific text
representation.}

\item{level_sep}{characters used to separate levels from models in
\code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{time_sep}{characters used to separate temporal blocks in
\code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{intercept}{characters used to represent the intercept in
\code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{intercept_sep}{characters used to the intercept from the other
parameters in \code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{open_p_value}{characters used as opening delimiters for the p value of
a node in \code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{close_p_value}{characters used as closing delimiters for the p value of
a node in \code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{open_model}{characters used as opening delimiters for the
representation of a model in \code{\link[=draw.covlmc]{draw.covlmc()}}.}

\item{close_model}{characters used as closing delimiters for the
representation of a model in \code{\link[=draw.covlmc]{draw.covlmc()}}.}
}
\value{
a list
}
\description{
This function returns a list of ASCII characters used to fine tune the \code{\link[=draw]{draw()}}
function behaviour when it is used with \code{format="text"}. It can be used
as is or customised using its parameters.
}
\examples{
charset_ascii(root = "x")

}
\seealso{
\code{\link[=draw]{draw()}}, \code{\link[=charset_utf8]{charset_utf8()}}.
}
