% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc_contexts.R
\name{contexts.covlmc}
\alias{contexts.covlmc}
\title{Contexts of a VLMC with covariates}
\usage{
\method{contexts}{covlmc}(ct)
}
\arguments{
\item{ct}{a fitted covlmc model.}
}
\value{
the different contexts present in the VLMC with covariates.
}
\description{
This function returns the different contexts present in a VLMC with covariates. Contexts' individual
states are given from the most recent to the least.
}
\examples{
pc <- powerconsumption[powerconsumption$week == 5, ]
breaks <- c(0, median(pc$active_power), max(pc$active_power))
labels <- c(0, 1)
dts <- cut(pc$active_power, breaks = breaks, labels = labels)
m_nocovariate <- vlmc(dts)
dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
m_cov <- covlmc(dts, dts_cov, min_size = 5)
contexts(m_cov)
}
