% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc.R
\name{cutoff}
\alias{cutoff}
\title{Cutoff values for pruning the context tree of a VLMC}
\usage{
cutoff(vlmc, mode = c("quantile", "native"), ...)
}
\arguments{
\item{vlmc}{a fitted VLMC model.}

\item{mode}{specify whether the results should be "native" likelihood ratio values
or expressed in a "quantile" scale of a chi-squared distribution (defaults to "quantile").}

\item{...}{additional arguments for the cutoff function.}
}
\value{
a vector of cut off values.
}
\description{
This function returns all the cutoff values that are guaranteed to induce a
pruning of the context tree of a VLMC.
}
