% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R
\name{draw}
\alias{draw}
\title{Text based representation of a context tree}
\usage{
draw(ct, terminals = FALSE, probs = FALSE)
}
\arguments{
\item{ct}{a context tree.}

\item{terminals}{not used}

\item{probs}{logical; if \code{TRUE}, display the next state distribution rather than the data counts}
}
\value{
the context tree (invisibly)
}
\description{
This function 'draws' a context tree as a text.
}
