% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_draw.R
\name{draw.vlmc}
\alias{draw.vlmc}
\title{Text based representation of a vlmc}
\usage{
\method{draw}{vlmc}(ct, control = draw_control(), node2txt = vlmc_node2txt, ...)
}
\arguments{
\item{ct}{a fitted vlmc.}

\item{control}{a list of low level control parameters of the text
representation. See details and \code{\link[=draw_control]{draw_control()}}.}

\item{node2txt}{an optional function called on each node to render it to a
text representation. Defaults to the frequencies of the states given the
context. Adding \code{prob=TRUE} as an additional parameter normalizes the
frequencies to probabilities.}

\item{...}{additional arguments for node2txt.}
}
\value{
the context tree (invisibly).
}
\description{
This function 'draws' a context tree as a text.
}
\details{
The function uses basic "ascii art" to represent the context tree. Characters
used to represent the structure of the tree, e.g. branches, can be modified
using \code{\link[=draw_control]{draw_control()}}.

If \code{node2txt} is non \code{NULL}, it is called for each node of the tree
(including leaves). If the result of the call is non \code{NULL}, it is included
in the text representation.
}
\examples{
dts <- sample(c("A", "B", "C"), 500, replace = TRUE)
model <- vlmc(dts, alpha = 0.05)
draw(model)
draw(model, prob = TRUE)
draw(model, node2txt = NULL)
}
