% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_draw2.R
\name{drawV}
\alias{drawV}
\title{Text based representation of a vlmc}
\usage{
drawV(ct, prob = TRUE)
}
\arguments{
\item{ct}{a fitted vlmc.}

\item{prob}{this parameter controls the display of node level information in
the tree. The default \code{prob=TRUE} represents the conditional distribution
of the states given the (partial) context associated to the node. Setting
\code{prob=FALSE} replaces the conditional distribution by the frequency of the
states that follow the context as in \code{\link[=draw.ctx_tree]{draw.ctx_tree()}}. Setting \code{prob=NULL}
removes all additional information.}
}
\description{
Text based representation of a vlmc
}
\examples{
dts <- sample(c("A", "B", "C"), 500, replace = TRUE)
model <- vlmc(dts, alpha = 0.05)
drawV(model)
drawV(model, prob = FALSE)
drawV(model, prob = NULL)
}
