% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R
\name{dts}
\alias{dts}
\title{Convert a vector to a discrete time series.}
\usage{
dts(x, vals = NULL)
}
\arguments{
\item{x}{a discrete time series; can be numeric, character, factor or
logical.}

\item{vals}{the set of values that can be taken by the time series, a.k.a. the
state space, see details (defaults to \code{NULL})}
}
\value{
a discrete time series (of class that inherits from \code{dts}).
}
\description{
This function creates a  representation of a discrete time series that
can be further processed by model estimation functions.
}
\details{
The discrete time series \code{x} can be a vector of numeric, character, factor or
logical type. If the state space of the series is not specified, that is when
\code{vals} is \code{NULL}, it is computed in a way that depends on the type of \code{x}:
\itemize{
\item for a factor, \code{vals} is set to the \code{levels()} of \code{x};
\item for a logical vector, \code{vals} is set to \code{c(FALSE, TRUE)};
\item for other types, \code{vals} is set to all the unique values taken by the time
series (as returned by \code{sort(unique(x))}).
}

If \code{vals} is specified, the function makes sure that \code{x} contains only the
specified values.
}
\examples{
x_dts <- dts(sample(c("A", "B"), 20, replace = TRUE))
x_dts
}
