% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R
\name{dts_data}
\alias{dts_data}
\title{Extract the plain discrete time series from a dts object}
\usage{
dts_data(x)
}
\arguments{
\item{x}{a dts object}
}
\value{
a vector representing the time seris
}
\description{
This function returns a copy of the discrete time series used to build the
dts object (see \code{\link[=dts]{dts()}}).
}
\examples{
raw_dts <- sample(c("A", "B", "C"), 50, replace = TRUE)
odts <- dts(raw_dts)
back_to_raw <- dts_data(odts)
## should be TRUE
identical(raw_dts, back_to_raw)
}
