% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_list.R
\name{dts_list}
\alias{dts_list}
\title{Convert a list of vectors/dts to a collection of discrete time series.}
\usage{
dts_list(xs, vals = NULL)
}
\arguments{
\item{xs}{a list of vectors or of \code{\link[=dts]{dts()}} objects that can be interpreted as discrete time series}

\item{vals}{the set of values that can be taken by the time series, a.k.a. the
state space, see details (defaults to \code{NULL})}
}
\value{
a collection of discrete time series, a list of class that inherits
from \code{dts_list}.
}
\description{
This function creates a representation of a collection of discrete time
series that can be further processed by model estimation functions.
}
\details{
The content of the list can be a collection of \code{\link[=dts]{dts()}} or of vectors.
The vectors of the list must all be of the same type and when interpreted
as discrete time series, they have to use a common state space. The possible
types are the one supported by \verb{[dts()]} (numeric, character, factor or
logical type).

If the state space of a collection of series is not specified, that is when
\code{vals} is \code{NULL}, it is computed base on the first vector \code{xs[[1]]} or
taken from the first \verb{[dts()]} object.

If \code{vals} is specified, the function makes sure that all time series use
only the allowed values.
}
\examples{
pre_dts <- lapply(1:20, \(x) sample(c("A", "B"), x + 20, replace = TRUE))
dts_collection <- dts_list(pre_dts, c("A", "B"))
dts_collection
## sub-collection
dts_collection[5:10]
## one time series
dts_collection[[15]]
}
\seealso{
\code{\link[=dts]{dts()}}
}
