% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_node_covlmc.R
\name{find_sequence.covlmc}
\alias{find_sequence.covlmc}
\title{Find the node of a sequence in a COVLMC context tree}
\usage{
\method{find_sequence}{covlmc}(ct, ctx, reverse = TRUE, ...)
}
\arguments{
\item{ct}{a context tree.}

\item{ctx}{a sequence to search in the context tree}

\item{reverse}{specifies whether the sequence \code{ctx} is given the reverse
temporal order (\code{TRUE}, default value) or in the temporal order (\code{FALSE}).}

\item{...}{additional parameters for the find_sequence function}
}
\value{
an object of class \code{ctx_node_covlmc} if the sequence \code{ctx} is represented
in the context tree, \code{NULL} when this is not the case
}
\description{
This function checks whether the sequence \code{ctx} is represented in the context
tree of the COVLMC model \code{ct}. If this is the case, it returns a description
of matching node, an object of class \code{ctx_node_covlmc}. If the sequence is
not represented in the tree, the function return \code{NULL}.
}
\details{
The function looks for sequences in general. The \code{\link[=is_context]{is_context()}} function can
be used on the resulting object to test if the sequence is in addition a
proper context.
}
\examples{
pc <- powerconsumption[powerconsumption$week == 5, ]
dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.5, 1))))
dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
m_cov <- covlmc(dts, dts_cov, min_size = 10)

## not in the tree
vals <- states(m_cov)
find_sequence(m_cov, c(vals[2], vals[2]), reverse = FALSE)
## in the tree but not a context
node <- find_sequence(m_cov, c(vals[1]), reverse = FALSE)
node
is_context(node)
## in the tree and a context
node <- find_sequence(m_cov, c(vals[1], vals[1]), reverse = FALSE)
node
is_context(node)
model(node)
}
