% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_node.R
\name{find_sequence}
\alias{find_sequence}
\alias{find_sequence.ctx_tree}
\title{Find the node of a sequence in a context tree}
\usage{
find_sequence(ct, ctx, reverse = TRUE, ...)

\method{find_sequence}{ctx_tree}(ct, ctx, reverse = TRUE, ...)
}
\arguments{
\item{ct}{a context tree.}

\item{ctx}{a sequence to search in the context tree}

\item{reverse}{specifies whether the sequence \code{ctx} is given the reverse
temporal order (\code{TRUE}, default value) or in the temporal order (\code{FALSE}).}

\item{...}{additional parameters for the find_sequence function}
}
\value{
an object of class \code{ctx_node} if the sequence \code{ctx} is represented
in the context tree, \code{NULL} when this is not the case.
}
\description{
This function checks whether the sequence \code{ctx} is represented in the context
tree \code{ct}. If this is the case, it returns a description of matching node, an
object of class \code{ctx_node}. If the sequence is not represented in the tree,
the function return \code{NULL}.
}
\details{
The function looks for sequences in general. The \code{\link[=is_context]{is_context()}} function can
be used on the resulting object to test if the sequence is in addition a
proper context.
}
\examples{
dts <- c("A", "B", "C", "A", "A", "B", "B", "C", "C", "A")
dts_tree <- ctx_tree(dts, max_depth = 3)
find_sequence(dts_tree, "A")
## returns NULL as "A" "C" does not appear in dts
find_sequence(dts_tree, c("A", "C"), reverse = FALSE)
}
