% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc.R
\name{is_covlmc}
\alias{is_covlmc}
\title{Test if the object is a covlmc model}
\usage{
is_covlmc(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} for VLMC models with covariates.
}
\description{
This function returns \code{TRUE} for VLMC models with covariates and \code{FALSE} for other objects.
}
