% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R
\name{is_ctx_tree}
\alias{is_ctx_tree}
\title{Test if the object is a context tree}
\usage{
is_ctx_tree(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} for context trees.
}
\description{
This function returns \code{TRUE} for context trees and \code{FALSE} for other objects.
}
