% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R
\name{is_dts}
\alias{is_dts}
\title{Test if the object is a discrete time series}
\usage{
is_dts(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} for discrete time series.
}
\description{
This function returns \code{TRUE} for discrete time series and \code{FALSE} for other objects.
}
\examples{
pre_dts <- sample(c("A", "B"), 20, replace = TRUE)
x_dts <- dts(pre_dts)
is_dts(x_dts)
is_dts(pre_dts)
}
