% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_list.R
\name{is_dts_list}
\alias{is_dts_list}
\title{Test if the object is a list of discrete time series}
\usage{
is_dts_list(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} for discrete time series list (as created by \code{\link[=dts_list]{dts_list()}})
}
\description{
This function returns \code{TRUE} for discrete time series list and \code{FALSE}
for other objects.
}
\examples{
pre_dts <- lapply(1:20, \(x) sample(c("A", "B"), x + 20, replace = TRUE))
dts_collection <- dts_list(pre_dts, c("A", "B"))
is_dts_list(dts_collection)
is_dts_list(pre_dts)
}
