% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_likelihood_multi.R
\name{loglikelihood.multi_vlmc}
\alias{loglikelihood.multi_vlmc}
\title{Log-Likelihood of a VLMC}
\usage{
\method{loglikelihood}{multi_vlmc}(
  vlmc,
  newdata,
  initial = c("truncated", "specific", "extended"),
  ignore,
  weights = NULL,
  ...
)
}
\arguments{
\item{vlmc}{the vlmc representation.}

\item{newdata}{either a discrete time series or a list of discrete time
series}

\item{initial}{specifies the likelihood function, more precisely the way the
first few observations for which contexts cannot be calculated are integrated
in the likelihood. Defaults to \code{"truncated"}. See below for details.}

\item{ignore}{specifies the number of initial values for which the loglikelihood
will not be computed. The minimal number depends on the likelihood function as
detailed below.}

\item{weights}{optional weights for the time series, see details.}

\item{...}{additional parameters for loglikelihood.}
}
\value{
an object of class \code{logLikMixVLMC} and \code{logLik}. This is a number,
the log-likelihood of the (CO)VLMC with the following attributes:
\itemize{
\item \code{df}: the number of parameters used by the VLMC for this likelihood calculation
\item \code{nobs}: the number of observations included in this likelihood calculation
\item \code{initial}: the value of the \code{initial} parameter used to compute this likelihood
}
}
\description{
This function evaluates the log-likelihood of a VLMC fitted on a
collection of discrete time series.
}
\details{
The definition of the likelihood function depends on the value of the
\code{initial} parameters, see the section below as well as the dedicated
vignette: \code{vignette("likelihood", package = "mixvlmc")}.

For VLMC objects, the method \code{loglikelihood.vlmc} will be used. For VLMC with covariables, \code{loglikelihood.covlmc}
will instead be called. For more informations on \code{loglikelihood} methods, use \code{methods(loglikelihood)} and their associated documentation.
}
\section{Limitation}{


VLMC fitted via \code{\link[=multi_vlmc]{multi_vlmc()}} on a collection discrete time series do not
support likelihood calculation without \code{newdata}.
}

\section{Weights}{


When specified, the weights are used to compute a weighted sum of the log
likelihood of the time series. The weights are interpreted as fractional
instances and the number of observations is therefore computed accordingly.
}

\section{likelihood calculation}{


In a (CO)VLMC of \code{\link[=depth]{depth()}}=k, we need k past values in order to compute the
context of a given observation. As a consequence, in a time series \code{x}, the
contexts of \code{x[1]} to \code{x[k]} are unknown. Depending on the value of \code{initial}
different likelihood functions are used to tackle this difficulty:
\itemize{
\item \code{initial=="truncated"}: the likelihood is computed using only
\code{x[(k+1):length(x)]}
\item \code{initial=="specific"}: the likelihood is computed on the full time series
using a specific context for the initial values, \code{x[1]} to \code{x[k]}. Each of
the specific context is unique, leading to a perfect likelihood of 1 (0 in
log scale). Thus the numerical value of the likelihood is identical as the
one obtained with \code{initial=="truncated"} but it is computed on \code{length(x)}
with a model with more parameters than in this previous case.
\item \code{initial=="extended"} (default): the likelihood is computed on the full time series
using an extended context matching for the initial values, \code{x[1]} to \code{x[k]}.
This can be seen as a compromised between the two other possibilities:
the relaxed context matching needs in general to turn internal nodes
of the context tree into actual context, increasing the number of parameters,
but not as much as with "specific". However, the likelihood of say \code{x[1]}
with an empty context is generally not 1 and thus the full likelihood is
smaller than the one computed with "specific".
}

In all cases, the \code{ignore} first values of the time series are not included
in the computed likelihood, but still used to compute contexts. If \code{ignore}
is not specified, it is set to the minimal possible value, that is k for the
\code{truncated} likelihood and 0 for the other ones. If it is specified, it must
be larger or equal to k for \code{truncated}.

See the dedicated vignette for a more mathematically oriented discussion:
\code{vignette("likelihood", package = "mixvlmc")}.
}

\examples{
pc <- powerconsumption[powerconsumption$week \%in\% 5:9, ]
powerlevels <- quantile(pc$active_power, probs = c(0.25, 0.5, 0.75, 1))
dts <- tapply(pc$active_power, pc$week, \(x) cut(x, breaks = c(0, powerlevels)))
model <- multi_vlmc(dts[-5], max_depth = 3)
loglikelihood(model, newdata = dts[[5]])
loglikelihood(model, newdata = dts[1:4])
}
\seealso{
\code{\link[=multi_vlmc]{multi_vlmc()}}
}
