% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree_multi.R
\name{multi_ctx_tree.default}
\alias{multi_ctx_tree.default}
\title{Build a context tree for a collection of discrete time series}
\usage{
\method{multi_ctx_tree}{default}(
  xs,
  min_size = 2L,
  max_depth = 100L,
  keep_position = FALSE,
  weights = NULL,
  ...
)
}
\arguments{
\item{xs}{a list of discrete times series}

\item{min_size}{positive numerical value (default: 2). Minimum number of observations for
a context to be included in the tree (counted over the full collection of
time series, see details for the case with \code{weights})}

\item{max_depth}{integer >= 1 (default: 100). Maximum length of a context to
be included in the tree.}

\item{keep_position}{logical (default: FALSE). Should the context tree keep
the position of the contexts.}

\item{weights}{optional weights for the time series, see details.}

\item{...}{additional parameters}
}
\value{
a context tree (of class that inherits from \code{multi_ctx_tree}).
}
\description{
This function builds a context tree for a collection of time series.
}
\details{
The tree represents all the sequences of symbols/states of length smaller
than \code{max_depth} that appear at least \code{min_size} times in collection of the
time series and stores the frequencies of the states that follow each
context.
}
\section{Weights}{


If given, the \code{weights} parameter must be a vector of non negative values of the
same length as \code{xs}. Each time series is then weighted using the corresponding
weight. Weights are interpreted as fractional number of occurrences when
\code{min_size} is checked. A context is kept in the context tree if the sum of
the weights of the series in which it appears is larger than the \code{min_size}
threshold.
}

\examples{
dts <- c(0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0)
dts2 <- c(0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0)
mdts <- list(dts, dts2)
mctx <- multi_ctx_tree(mdts, max_depth = 4)
}
\seealso{
\code{\link[=ctx_tree]{ctx_tree()}}, \code{\link[=dts_list]{dts_list()}}
}
