% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R
\name{print.dts}
\alias{print.dts}
\title{Print a discrete time series}
\usage{
\method{print}{dts}(x, n = 5, ...)
}
\arguments{
\item{x}{the \code{dts} object to print}

\item{n}{the number of time steps of time series to print (defaults to 5)}

\item{...}{additional arguments for the print function.}
}
\value{
the \code{x} object, invisibly
}
\description{
This function prints a discrete time series.
}
\examples{
x_dts <- dts(sample(c("A", "B"), 20, replace = TRUE))
print(x_dts, n = 10)
}
