% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmlc.R
\name{vlmc}
\alias{vlmc}
\title{Fit a Variable Length Markov Chain (VLMC)}
\usage{
vlmc(x, alpha = 0.05, min_size = 2, max_depth = 100)
}
\arguments{
\item{x}{a discrete time series; can be numeric, character or factor.}

\item{alpha}{number in (0,1) (default: 0.05) cut off value in the pruning phase.}

\item{min_size}{integer >= 1 (default: 2). Minimum number of observations for
a context in the growing phase of the context tree.}

\item{max_depth}{integer >= 1 (default: 100). Longest context considered in
growing phase of the context tree.}
}
\value{
a fitted vlmc model
}
\description{
This function fits a  Variable Length Markov Chain (VLMC) to a discrete time series.
}
