% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc_as.R
\name{as_covlmc}
\alias{as_covlmc}
\title{Convert an object to a Variable Length Markov Chain with covariates (coVLMC)}
\usage{
as_covlmc(x, ...)
}
\arguments{
\item{x}{an object to convert into a covlmc.}

\item{...}{additional arguments for conversion functions.}
}
\value{
a covlmc
}
\description{
This generic function converts an object into a covlmc.
}
