% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc_contexts.R
\name{contexts.covlmc}
\alias{contexts.covlmc}
\title{Contexts of a VLMC with covariates}
\usage{
\method{contexts}{covlmc}(ct, type = c("list", "data.frame"), reverse = FALSE, ...)
}
\arguments{
\item{ct}{a fitted covlmc model.}

\item{type}{result type (see details).}

\item{reverse}{logical (defaults to FALSE). See details.}

\item{...}{additional arguments for the contexts function.}
}
\value{
the list of the contexts represented in this tree or a data.frame
  with more content.
}
\description{
This function returns the different contexts present in a VLMC with covariates.
}
\examples{
pc <- powerconsumption[powerconsumption$week == 5, ]
breaks <- c(0, median(pc$active_power), max(pc$active_power))
labels <- c(0, 1)
dts <- cut(pc$active_power, breaks = breaks, labels = labels)
m_nocovariate <- vlmc(dts)
dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
m_cov <- covlmc(dts, dts_cov, min_size = 5)
contexts(m_cov)
}
