% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree_contexts.R
\name{contexts}
\alias{contexts}
\title{Contexts of a context tree}
\usage{
contexts(ct)
}
\arguments{
\item{ct}{a context tree.}
}
\value{
the list of the contexts represented in this tree.
}
\description{
This function extracts from a context tree the list of all its contexts.
}
