% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc.R
\name{cutoff.covlmc}
\alias{cutoff.covlmc}
\title{Cutoff values for pruning the context tree of a VLMC with covariates}
\usage{
\method{cutoff}{covlmc}(vlmc, mode = c("quantile", "native"), ...)
}
\arguments{
\item{vlmc}{a fitted covlmc model}

\item{mode}{specify whether the results should be "native" likelihood ratio
values or expressed in a "quantile" scale of a chi-squared distribution.
For covlmc, only the quantile scale is supported.}

\item{...}{additional arguments for the cutoff function}
}
\value{
a vector of cut off values
}
\description{
This function returns all the cutoff values that are guaranteed to induce a
pruning of the context tree of a VLMC with covariates
}
\examples{
pc <- powerconsumption[powerconsumption$week == 5, ]
dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.5, 1))))
m_nocovariate <- vlmc(dts)
draw(m_nocovariate)
dts_cov <- data.frame(day_night = (pc$hour>=7 & pc$hour<=17))
m_cov <- covlmc(dts, dts_cov, min_size = 5)
draw(m_cov)
}
