% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc.R
\name{cutoff}
\alias{cutoff}
\title{Cutoff values for pruning the context tree of a VLMC}
\usage{
cutoff(vlmc, mode = c("quantile", "native"), raw = FALSE, ...)
}
\arguments{
\item{vlmc}{a fitted VLMC model.}

\item{mode}{specify whether the results should be "native" likelihood ratio values
or expressed in a "quantile" scale of a chi-squared distribution (defaults to "quantile").}

\item{raw}{specify whether the returned values should be limit values computed in the model or
modified values that guarantee pruning (see details)}

\item{...}{additional arguments for the cutoff function.}
}
\value{
a vector of cut off values.
}
\description{
This function returns all the cutoff values that are guaranteed to induce a
pruning of the context tree of a VLMC. Pruning is implemented by the \code{\link[=prune]{prune()}} function.
}
\details{
By default, the function tries to return values that can be used directly to
induce pruning in the context tree. This is done by returning values that are
slightly smaller than values computed in the context tree in the "quantile" scale
and slightly larger in the "native" scale. However, this approach does not allow
full consistency between scales: while the cutoff values correspond when transformed
from one scale to another, round off errors could induce differences in the results.

Setting the \code{raw} parameter to \code{TRUE} removes this operation on the values and shifts
the burden of value perturbation to the user.
}
\examples{
pc <- powerconsumption[powerconsumption$week == 5, ]
dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.25, 0.5, 0.75, 1))))
model <- vlmc(dts)
draw(model)
model_cuts <- cutoff(model)
model_2 <- prune(model, model_cuts[2])
draw(model_2)
}
\seealso{
\code{\link[=prune]{prune()}}
}
