% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R
\name{depth}
\alias{depth}
\title{Depth of a context tree}
\usage{
depth(ct)
}
\arguments{
\item{ct}{a context tree.}
}
\value{
the depth of the tree.
}
\description{
This function return the depth of a context tree, i.e. the length of the
longest context represented in the tree.
}
\examples{
dts <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 0)
dts_ctree <- ctx_tree(dts, min_size = 1, max_depth = 3)
# should be 3
depth(dts_ctree)
}
