% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc_draw.R
\name{draw.covlmc}
\alias{draw.covlmc}
\title{Text based representation of a covlmc model}
\usage{
\method{draw}{covlmc}(ct, node2txt = draw_covlmc_node, ...)
}
\arguments{
\item{ct}{a fitted covlmc model.}

\item{node2txt}{an optional function called on each node to render it to a text representation}

\item{...}{additional arguments for node2txt (see details)}
}
\value{
the covlmc model (invisibly)
}
\description{
This function 'draws' a covlmc model as a text.
}
\details{
Numerical parameters and p-values are represented using the [base::signif] function.
The number of significant digits can be specified using a `digits` parameter.
}
