% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree_draw.R
\name{draw}
\alias{draw}
\title{Text based representation of a context tree}
\usage{
draw(ct, node2txt = NULL, ...)
}
\arguments{
\item{ct}{a context tree.}

\item{node2txt}{an optional function called on each node to render it to a text representation.}

\item{...}{additional arguments for node2txt.}
}
\value{
the context tree (invisibly).
}
\description{
This function 'draws' a context tree as a text.
}
