% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R
\name{is_ctx_tree}
\alias{is_ctx_tree}
\title{Test if the object is a context tree}
\usage{
is_ctx_tree(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} for context trees.
}
\description{
This function returns \code{TRUE} for context trees and \code{FALSE} for other objects.
}
\examples{
rdts <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 0)
rdts_ctree <- ctx_tree(rdts, min_size = 1, max_depth = 2)
is_ctx_tree(rdts_ctree)
is_ctx_tree(rdts)
}
