% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc.R
\name{is_vlmc}
\alias{is_vlmc}
\title{Test if the object is a vlmc model}
\usage{
is_vlmc(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
\code{TRUE} for VLMC models
}
\description{
This function returns \code{TRUE} for VLMC models and \code{FALSE} for other objects.
}
