% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_likelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log-Likelihood of a VLMC}
\usage{
loglikelihood(vlmc, x = NULL)
}
\arguments{
\item{vlmc}{the vlmc representation}

\item{x}{an optional discrete time series}
}
\value{
the log-likelihood of the VLMC
}
\description{
This function evaluates the log-likelihood of a VLMC fitted on a discrete time series.
When the optional argument \code{x} is provided, the function evaluates instead the
log-likelihood for this (new) discrete time series.
}
\seealso{
[stats::logLik]
}
