% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree_multi.R
\name{multi_ctx_tree}
\alias{multi_ctx_tree}
\title{Build a context tree for a collection of discrete time series}
\usage{
multi_ctx_tree(xs, min_size = 2L, max_depth = 25L, keep_position = FALSE)
}
\arguments{
\item{xs}{a list of discrete times series}

\item{min_size}{integer >= 1 (default: 2). Minimum number of observations for
a context to be included in the tree (counted over the full collection of
time series, see details)}

\item{max_depth}{integer >= 1 (default: 25). Maximum length of a context to
be included in the tree.}

\item{keep_position}{logical (default: FALSE). Should the context tree keep
the position of the contexts.}
}
\value{
a context tree (of class that inherits from \code{multi_ctx_tree}).
}
\description{
This function builds a context tree for a collection of time series.
}
\details{
The tree represents all the sequences of symbols/states of length smaller
than \code{max_depth} that appear at least \code{min_size} times in collection of the
time series and stores the frequencies of the states that follow each
context.

Owing to the iterative nature of construction, this function may use a large
quantity of memory as pruning infrequent contexts is only done after
computing all of them. It is therefore recommend to avoid large depths and
the default value of \code{max_depth} is smaller than in the single time series
function \code{\link[=ctx_tree]{ctx_tree()}}.
}
\examples{
dts <- c(0, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 0)
dts2 <- c(0, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0)
mdts <- list(dts, dts2)
mctx <- multi_ctx_tree(mdts, max_depth = 4)
}
\seealso{
\code{\link[=ctx_tree]{ctx_tree()}}
}
