% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc_predict.R
\name{predict.vlmc}
\alias{predict.vlmc}
\title{Predictor of the following state of a discrete time series for a vlmc}
\usage{
\method{predict}{vlmc}(object, ts, type = "class", ...)
}
\arguments{
\item{object}{a fitted vlmc object.}

\item{ts}{a time series adapted to the vlmc object.}

\item{type}{charachter indicating the type of prediction required.
The default "class" returns the predictor,
"probs" returns a matrix of prediction probabilities.}

\item{...}{additional arguments.}
}
\description{
This function predict each state of a time series from the distribution estimated by the
given vlmc object.
}
\examples{
pc <- powerconsumption[powerconsumption$week == 5, ]
dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.25, 0.5, 0.75, 1))))
model <- vlmc(dts, min_size = 5)
predict(model, dts[1:5])
}
