% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts.R
\name{print.contexts}
\alias{print.contexts}
\title{Print a context list}
\usage{
\method{print}{contexts}(x, reverse = TRUE, ...)
}
\arguments{
\item{x}{the \code{contexts} object to print}

\item{reverse}{specifies whether the contexts should be reported in reverse
temporal order (\code{TRUE}, default value) or in the temporal order (\code{FALSE}).}

\item{...}{additional arguments for the print function.}
}
\value{
the \code{x} object, invisibly
}
\description{
This function prints a list of contexts i.e. a \code{contexts} object listing
\code{ctx_node} objects.
}
\examples{
dts <- c("A", "B", "C", "A", "A", "B", "B", "C", "C", "A")
dts_tree <- ctx_tree(dts, max_depth = 3)
print(contexts(dts_tree))
}
\seealso{
\code{\link[=contexts]{contexts()}}
}
