% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmlc.R
\name{prune}
\alias{prune}
\title{Prune a Variable Length Markov Chain (VLMC)}
\usage{
prune(vlmc, alpha = 0.05, cutoff = NULL, ...)
}
\arguments{
\item{vlmc}{a fitted VLMC model}

\item{alpha}{number in (0,1) (default: 0.05) cutoff value in quantile scale for pruning}

\item{cutoff}{positive number: cutoff value in native (likelihood ratio) scale for pruning.
Defaults to the value obtained from \code{alpha}. Takes precedence over \code{alpha} is specified.}

\item{...}{additional arguments for the prune function}
}
\value{
a pruned VLMC
}
\description{
This function prunes a VLMC
}
