% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlmc_simulate.R
\name{simulate.covlmc}
\alias{simulate.covlmc}
\title{Simulate a discrete time series}
\usage{
\method{simulate}{covlmc}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{a fitted covlmc object.}

\item{nsim}{length of the simulated time series (defaults to 1).}

\item{seed}{an optional random seed.}

\item{...}{additional arguments (see details).}
}
\description{
This function simulates a time series from the distribution estimated by the
given covlmc object.
}
\details{
A VLMC with covariates model needs covariates to compute its transition
probabilities. The covariates must be submitted as a data frame using the
named \code{covariate} argument.
}
