% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R, R/dts.R, R/dts_list.R
\name{states}
\alias{states}
\alias{states.ctx_tree}
\alias{states.dts}
\alias{states.dts_list}
\title{State space of an object}
\usage{
states(x)

\method{states}{ctx_tree}(x)

\method{states}{dts}(x)

\method{states}{dts_list}(x)
}
\arguments{
\item{x}{an object with a state space.}
}
\value{
the state space of the object (as vector of states).
}
\description{
This function returns the state space of an object for which this is
meaningful such as a discrete time series or a context tree.
}
\examples{
rdts <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 0)
rdts_ctree <- ctx_tree(rdts, min_size = 1, max_depth = 2)
## should be c(0, 1)
states(rdts_ctree)
x_dts <- dts(sample(c("A", "B", "C"), 20, replace = TRUE))
## should be c("A", "B", "C")
states(x_dts)
pre_dts <- lapply(1:10, \(x) sample(c("A", "B", "C"), x + 10, replace = TRUE))
dts_collection <- dts_list(pre_dts, c("A", "B", "C"))
states(dts_collection)
}
