% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R, R/dts.R
\name{states}
\alias{states}
\alias{states.ctx_tree}
\alias{states.dts}
\title{State space of an object}
\usage{
states(x)

\method{states}{ctx_tree}(x)

\method{states}{dts}(x)
}
\arguments{
\item{x}{an object with a state space.}
}
\value{
the state space of the context tree.
}
\description{
This function returns the state space of an object for which this is
meaningful such as a discrete time series or a context tree.
}
\examples{
rdts <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 0)
rdts_ctree <- ctx_tree(rdts, min_size = 1, max_depth = 2)
## should be c(0, 1)
states(rdts_ctree)
x_dts <- dts(sample(c("A", "B", "C"), 20, replace = TRUE))
## should be c("A", "B", "C")
states(x_dts)
}
