% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctx_tree.R
\name{states}
\alias{states}
\title{State space of a context tree}
\usage{
states(ct)
}
\arguments{
\item{ct}{a context tree.}
}
\value{
the context space of the tree.
}
\description{
This function returns the state space of a context tree.
}
\examples{
dts <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 0)
dts_ctree <- ctx_tree(dts, min_size = 1, max_depth = 2)
## should be c("0", "1")
states(dts_ctree)
}
