% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlmc.R
\name{vlmc}
\alias{vlmc}
\title{Fit a Variable Length Markov Chain (VLMC)}
\usage{
vlmc(
  x,
  alpha = 0.05,
  cutoff = NULL,
  min_size = 2,
  max_depth = 100,
  prune = TRUE
)
}
\arguments{
\item{x}{a discrete time series; can be numeric, character or factor.}

\item{alpha}{number in (0,1) (default: 0.05) cutoff value in quantile scale in the pruning phase.}

\item{cutoff}{positive number: cutoff value in native (likelihood ratio) scale in the pruning phase.
Defaults to the value obtained from \code{alpha}. Takes precedence over \code{alpha} is specified.}

\item{min_size}{integer >= 1 (default: 2). Minimum number of observations for
a context in the growing phase of the context tree.}

\item{max_depth}{integer >= 1 (default: 100). Longest context considered in
growing phase of the context tree.}

\item{prune}{logical: specified whether the context tree should be pruned (default behavior).}
}
\value{
a fitted vlmc model.
}
\description{
This function fits a  Variable Length Markov Chain (VLMC) to a discrete time series.
}
